\name{TMCoCitSupportFM-package}
\alias{TMCoCitSupportFM-package}
\alias{TMCoCitSupportFM}
\docType{package}
\title{
\packageTitle{TMCoCitSupportFM}
}
\description{
\packageDescription{TMCoCitSupportFM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TMCoCitSupportFM}
\packageIndices{TMCoCitSupportFM}
Use this Package in conjounting with a Co Citation Analysis and Web of Knowledge exported Abstracts
}
\author{
\packageAuthor{TMCoCitSupportFM}

Maintainer: Max Finsterbusch \packageMaintainer{TMCoCitSupportFM}
}
\references{

}

\keyword{ package }
\seealso{


}
\examples{

}
